[Modify] 대략적 전투 수정

+++ Scripts/Action/Dungeon/Skill.cs
@@ -92,7 +92,7 @@ namespace Nightmare
                     //공격 -> 이름/데미지/타겟/소모마나/쿨타임/지속시간/누가하냐(보스2)/얼마나 자주 때리냐
                     //스텟 -> 이름/데미지/타겟/소모마나/쿨타임/어떤거/어디/지속 시간/보스냐 플레이어냐
                     case 1:
-                        Skill Snowwhite1 = new AttackSkill("숲의 공주님", (boss.MonsterAttack)*(20/100), 1, 10, 3, 0, 2,2);
+                        Skill Snowwhite1 = new AttackSkill("숲의 공주님", (float)((boss.MonsterAttack)*(20.0/100)), 1, 10, 3, 0, 2,2);
                         Skill Snowwhite2 = new StatSkill("새들의 합창", (boss.MonsterHealth)*(15/100), 0, 0, 3, "버프", "체력", 0, 2);
                         Skill Snowwhite3 = new StatSkill("자연의 원망", 5, 1, 0, 3, "디버프", "회피율", 3, 2);
                         boss.BossSkill.Add(Snowwhite1);
@@ -100,23 +100,23 @@ namespace Nightmare
                         boss.BossSkill.Add(Snowwhite3);
                         break;
                     case 2:
-                        Skill Cinderella1 = new AttackSkill("재투성이", (boss.MonsterAttack) * (30 / 100),1,0,3,0,2,1);
+                        Skill Cinderella1 = new AttackSkill("재투성이", (float)((boss.MonsterAttack) * (30.0 / 100)), 1,0,3,0,2,1);
                         Skill Cinderella2 = new StatSkill("신분 상승", boss.MonsterAttack, 0, 0, 3, "버프", "체력", 3, 2);
-                        Skill Cinderella3 = new StatSkill("대모의 응원", (player.Stat.BaseDef+player.Stat.EquipDef) * (20/100), 0, 0, 3, "디버프", "방어력", 3, 2);
+                        Skill Cinderella3 = new StatSkill("대모의 응원", (float)((player.Stat.BaseDef+player.Stat.EquipDef) * (20.0/100)), 0, 0, 3, "디버프", "방어력", 3, 2);
                         boss.BossSkill.Add(Cinderella1);
                         boss.BossSkill.Add(Cinderella2);
                         boss.BossSkill.Add(Cinderella3);
                         break;
                     case 3:
-                        Skill aurora1 = new AttackSkill("오로라", (boss.MonsterAttack) * (50 / 100), 1, 0, 3, 0, 2,1);
+                        Skill aurora1 = new AttackSkill("오로라", (float)((boss.MonsterAttack) * (50.0 / 100)), 1, 0, 3, 0, 2,1);
                         Skill aurora2 = new StatSkill("백 년 가약", 10, 0, 0, 3, "버프", "치명타율", 3, 2);
-                        Skill aurora3 = new StatSkill("운명의 소용돌이", (player.Stat.EquipAtk+player.Stat.BaseAtk)*(10/100), 0, 0, 3, "디버프", "공격력", 3, 2);
+                        Skill aurora3 = new StatSkill("운명의 소용돌이", (float)((player.Stat.EquipAtk+player.Stat.BaseAtk)*(10.0/100)), 0, 0, 3, "디버프", "공격력", 3, 2);
                         boss.BossSkill.Add(aurora1);
                         boss.BossSkill.Add(aurora2);
                         boss.BossSkill.Add(aurora3);
                         break;
                     case 4:
-                        Skill Arial1 = new AttackSkill("인간 갈망", (boss.MonsterAttack) * (15 / 100), 2, 10, 3, 0, 2,3);
+                        Skill Arial1 = new AttackSkill("인간 갈망", (float)((boss.MonsterAttack) * (15.0 / 100)), 2, 10, 3, 0, 2,3);
                         Skill Arial2 = new StatSkill("현혹의 노래", boss.MonsterDefense, 1, 10, 3, "버프", "체력", 0, 2);
                         Skill Arial3 = new StatSkill("비극적인 결말", 10, 1, 10, 3, "디버프", "마력", 0, 2);
                         boss.BossSkill.Add(Arial1);
@@ -124,8 +124,8 @@ namespace Nightmare
                         boss.BossSkill.Add(Arial3);
                         break;
                     case 5:
-                        Skill Belle1 = new AttackSkill("괴물 혐오", (boss.MonsterAttack) * (30 / 100), 1, 10, 3, 0, 2,2);
-                        Skill Belle2 = new StatSkill("마음의 식량", (boss.MonsterHealth) * (20 / 100), 0, 0,3,"버프", "체력", 0, 2);
+                        Skill Belle1 = new AttackSkill("괴물 혐오", (float)((boss.MonsterAttack) * (30.0 / 100)), 1, 10, 3, 0, 2,2);
+                        Skill Belle2 = new StatSkill("마음의 식량", (float)((boss.MonsterHealth) * (20.0 / 100)), 0, 0,3,"버프", "체력", 0, 2);
                         Skill Belle3 = new StatSkill("사냥꾼의 정신", 7, 0, 0, 3, "디버프", "치명타율", 3, 2);
                         boss.BossSkill.Add(Belle1);
                         boss.BossSkill.Add(Belle2);
